refactor: RecommendField 컴포넌트 분리
